Undo action at doc modification, deletion disables editor (no loss of data)

+++ Forms/PythonScriptForm.cs
@@ -9,6 +9,8 @@ using Grasshopper.Kernel.Types;
 using GhPython.Component;
 using System.IO;
 using System.Diagnostics;
+using System.Reflection;
+using Grasshopper.Kernel.Undo;
 
 namespace GhPython.Forms
 {
@@ -16,34 +18,48 @@ namespace GhPython.Forms
     {
         Control _texteditor;
         bool _showClosePrompt = true;
+
+        /// <summary>
+        /// The linked component. This field might be null.
+        /// </summary>
         PythonComponent _component;
+        int _targetVariableMenuIndex = -1;
 
         // keep default constructor around to not "confuse" Visual Studio's designer
-        public PythonScriptForm()
+        public PythonScriptForm() : this(null)
         {
-            InitializeComponent();
         }
         public PythonScriptForm(PythonComponent linkedComponent)
         {
             InitializeComponent();
-            this.KeyDown += new KeyEventHandler(ScriptForm_KeyDown);
+            this.KeyDown += ScriptForm_KeyDown;
 
             _component = linkedComponent;
-            _texteditor = _component.CreateEditorControl(OnPythonHelp);
-            this.splitContainer1.Panel1.Controls.Add(_texteditor);
-            _texteditor.Dock = DockStyle.Fill;
 
-            var inputCode = _component.Params.Input[0];
-            if (inputCode.VolatileDataCount > 0 && inputCode.VolatileData.PathCount > 0)
+            if (_component != null)
             {
-                var goo = inputCode.VolatileData.get_Branch(0)[0] as IGH_Goo;
-                if (goo != null)
+                _texteditor = _component.CreateEditorControl(OnPythonHelp);
+                this.splitContainer.Panel1.Controls.Add(_texteditor);
+                _texteditor.Dock = DockStyle.Fill;
+
+                var inputCode = _component.Params.Input[0];
+                if (inputCode.VolatileDataCount > 0 && inputCode.VolatileData.PathCount > 0)
                 {
-                    string code;
-                    if (goo.CastTo(out code) && !string.IsNullOrEmpty(code))
-                        _texteditor.Text = code;
+                    var goo = inputCode.VolatileData.get_Branch(0)[0] as IGH_Goo;
+                    if (goo != null)
+                    {
+                        string code;
+                        if (goo.CastTo(out code) && !string.IsNullOrEmpty(code))
+                            _texteditor.Text = code;
+                    }
                 }
+
+
+                _targetVariableMenuIndex = 
+                    fileToolStripMenuItem.DropDownItems.Add(_component.GetTargetVariableMenuItem());
             }
+
+            versionLabel.Text = Assembly.GetExecutingAssembly().GetName().Version.ToString();
         }
 
         void OnPythonHelp(string str)
@@ -101,38 +117,46 @@ namespace GhPython.Forms
 
         private void SetDefinitionValue(bool close, bool expire)
         {
-            IList<IGH_Param> outs = _component.Params.Input;
-
-            if (outs.Count > 0)
+            if (_component != null)
             {
-                var codeInput = outs[0] as GH_PersistentParam<GH_String>;
+                IList<IGH_Param> outs = _component.Params.Input;
 
-                if (codeInput != null)
+                if (outs.Count > 0)
                 {
-                    if (codeInput.SourceCount != 0)
+                    var codeInput = outs[0] as GH_PersistentParam<GH_String>;
+
+                    if (codeInput != null)
                     {
-                        if (MessageBox.Show("There is dynamic inherited input that overrides this components behaviour.\nPlease unlink the first input to see the result.",
-                            "Rhino.Python", MessageBoxButtons.OKCancel) == DialogResult.Cancel)
+                        if (codeInput.SourceCount != 0)
                         {
-                            return;
+                            if (MessageBox.Show("There is dynamic inherited input that overrides this components behaviour.\nPlease unlink the first input to see the result.",
+                                "Rhino.Python", MessageBoxButtons.OKCancel) == DialogResult.Cancel)
+                            {
+                                return;
+                            }
                         }
-                    }
 
-                    string newCode = _texteditor.Text;
+                        GH_Document ghd = _component.OnPingDocument();
+                        if(ghd != null)
+                        ghd.UndoServer.PushUndoRecord("Python code changed",
+                            new Grasshopper.Kernel.Undo.Actions.GH_GenericObjectAction(codeInput));
 
-                    codeInput.ClearPersistentData();
-                    if (!string.IsNullOrEmpty(newCode))
-                    {
-                        codeInput.AddPersistentData(new GH_String(newCode));
-                    }
+                        codeInput.ClearPersistentData();
+                        string newCode = _texteditor.Text;
 
-                    if(expire)
-                        codeInput.ExpireSolution(true);
+                        if (!string.IsNullOrEmpty(newCode))
+                        {
+                            codeInput.AddPersistentData(new GH_String(newCode));
+                        }
 
-                    if (close)
-                    {
-                        _showClosePrompt = false;
-                        Close();
+                        if (expire)
+                            codeInput.ExpireSolution(true);
+
+                        if (close)
+                        {
+                            _showClosePrompt = false;
+                            Close();
+                        }
                     }
                 }
             }
@@ -169,21 +193,36 @@ namespace GhPython.Forms
 
             if (!e.Cancel)
             {
-                var attributes = _component.Attributes as PythonComponentAttributes;
-
-                if (attributes != null)
+                if (_component != null)
                 {
-                    attributes.ClearLinkedForm(false);
+                    var attributes = _component.Attributes as PythonComponentAttributes;
+
+                    if (attributes != null)
+                    {
+                        attributes.DisableLinkedForm(false);
+                    }
                 }
             }
             
             base.OnClosing(e);
         }
 
-        public void CloseWithoutUserCancel()
+        public void Disable()
         {
             _showClosePrompt = false;
-            Close();
+
+            SuspendLayout();
+            okButton.Enabled = false;
+            testButton.Enabled = false;
+
+            this.Text += " (disabled)";
+            mainStatusText.Text = "Window is disabled because linked component was cancelled.";
+
+            if(_targetVariableMenuIndex >= 0 && _targetVariableMenuIndex < fileToolStripMenuItem.DropDownItems.Count)
+                fileToolStripMenuItem.DropDownItems.RemoveAt(_targetVariableMenuIndex);
+
+            _component = null;
+            ResumeLayout(true);
         }
 
         private void LastHandleException(Exception ex)
@@ -195,7 +234,7 @@ namespace GhPython.Forms
             }
         }
 
-        private void openToolStripMenuItem_Click(object sender, EventArgs e)
+        private void importFrom_Click(object sender, EventArgs e)
         {
             try
             {
@@ -217,7 +256,7 @@ namespace GhPython.Forms
             }
         }
 
-        private void saveToolStripMenuItem_Click(object sender, EventArgs e)
+        private void exportAs_Click(object sender, EventArgs e)
         {
             try
             {
@@ -255,6 +294,30 @@ namespace GhPython.Forms
             try
             {
                 Process.Start("http://www.grasshopper3d.com/forum/");
+            }
+            catch (Exception ex)
+            {
+                LastHandleException(ex);
+            }
+        }
+
+        private void pythonDocumentationToolStripMenuItem_Click(object sender, EventArgs e)
+        {
+            try
+            {
+                Process.Start("http://docs.python.org/");
+            }
+            catch (Exception ex)
+            {
+                LastHandleException(ex);
+            }
+        }
+
+        private void closeToolStripMenuItem_Click(object sender, EventArgs e)
+        {
+            try
+            {
+                Close();
             }
             catch (Exception ex)
             {
